requirements.txt
# Core dependencies for topic modeling
bertopic>=0.15.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
tqdm>=4.65.0

# Data collection (scraping) dependencies
aiohttp>=3.8.0
python-dotenv>=1.0.0
